+++ exporter/exporterhelper/common.go
@@ -17,19 +17,6 @@ import (
 	"go.opentelemetry.io/collector/obsreport"
 )
 
-// TimeoutSettings for timeout. The timeout applies to individual attempts to send data to the backend.
-type TimeoutSettings struct {
-	// Timeout is the timeout for every attempt to send data to the backend.
-	Timeout time.Duration `mapstructure:"timeout"`
-}
-
-// NewDefaultTimeoutSettings returns the default settings for TimeoutSettings.
-func NewDefaultTimeoutSettings() TimeoutSettings {
-	return TimeoutSettings{
-		Timeout: 5 * time.Second,
-	}
-}
-
 // requestSender is an abstraction of a sender for a request independent of the type of the data (traces, metrics, logs).
 type requestSender interface {
 	start(ctx context.Context, host component.Host, set exporter.CreateSettings) error
@@ -249,24 +236,6 @@ func (be *baseExporter) setOnTemporaryFailure(onTemporaryFailure onRequestHandli
 	if rs, ok := be.retrySender.(*retrySender); ok {
 		rs.onTemporaryFailure = onTemporaryFailure
 	}
-}
-
-// timeoutSender is a requestSender that adds a `timeout` to every request that passes this sender.
-type timeoutSender struct {
-	baseRequestSender
-	cfg TimeoutSettings
-}
-
-func (ts *timeoutSender) send(req internal.Request) error {
-	// Intentionally don't overwrite the context inside the request, because in case of retries deadline will not be
-	// updated because this deadline most likely is before the next one.
-	ctx := req.Context()
-	if ts.cfg.Timeout > 0 {
-		var cancelFunc func()
-		ctx, cancelFunc = context.WithTimeout(req.Context(), ts.cfg.Timeout)
-		defer cancelFunc()
-	}
-	return req.Export(ctx)
 }
 
 func createSampledLogger(logger *zap.Logger) *zap.Logger {
